Display all public recipe lis

+++ db/migrate/20221213180858_create_foods.rb
@@ -3,7 +3,7 @@ class CreateFoods < ActiveRecord::Migration[7.0]
     create_table :foods do |t|
       t.string :name
       t.string :measurement_unit
-      t.decimal :price
+      t.integer :price
       t.integer :quantity
       t.references :user, null: false, foreign_key: true
